--- a/src/systems/jumps/jumpsystem.jl
+++ b/src/systems/jumps/jumpsystem.jl
@@ -141,12 +141,17 @@ dprob = DiscreteProblem(js, u₀map, tspan, parammap)
 """
 function DiffEqBase.DiscreteProblem(sys::AbstractSystem, u0map, tspan::Tuple,
                                     parammap=DiffEqBase.NullParameters(); kwargs...)
-    #u0 = varmap_to_vars(u0map, states(sys))
-    #p  = varmap_to_vars(parammap, parameters(sys))
+
+    (u0map isa AbstractVector) || error("For DiscreteProblems u0map must be an AbstractVector.")
     u0d = Dict( convert(Variable,u[1]) => u[2] for u in u0map)
-    pd  = Dict( convert(Variable,u[1]) => u[2] for u in parammap)
     u0 = [u0d[u] for u in states(sys)]
-    p  = [pd[u] for u in parameters(sys)]
+    if parammap != DiffEqBase.NullParameters()
+        (parammap isa AbstractVector) || error("For DiscreteProblems parammap must be an AbstractVector.")
+        pd  = Dict( convert(Variable,u[1]) => u[2] for u in parammap)    
+        p  = [pd[u] for u in parameters(sys)]
+    else
+        p = parammap
+    end
     f  = (du,u,p,t) -> du.=u    # identity function to make syms works
     df = DiscreteFunction(f, syms=Symbol.(states(sys)))
     DiscreteProblem(df, u0, tspan, p; kwargs...)