--- a/src/systems/jumps/jumpsystem.jl
+++ b/src/systems/jumps/jumpsystem.jl
@@ -27,7 +27,7 @@ js      = JumpSystem([j₁,j₂,j₃], t, [S,I,R], [β,γ])
 """
 struct JumpSystem{U <: ArrayPartition} <: AbstractSystem
     """
-    The jumps of the system. Allowable types are `ConstantRateJump`, 
+    The jumps of the system. Allowable types are `ConstantRateJump`,
     `VariableRateJump`, `MassActionJump`.
     """
     eqs::U
@@ -48,7 +48,7 @@ function JumpSystem(eqs, iv, states, ps; systems = JumpSystem[],
 
     ap = ArrayPartition(MassActionJump[], ConstantRateJump[], VariableRateJump[])
     for eq in eqs
-        if eq isa MassActionJump 
+        if eq isa MassActionJump
             push!(ap.x[1], eq)
         elseif eq isa ConstantRateJump
             push!(ap.x[2], eq)
@@ -73,7 +73,7 @@ generate_affect_function(js, affect, outputidxs) = build_function(affect, states
                                                       expression=Val{false},
                                                       headerfun=add_integrator_header,
                                                       outputidxs=outputidxs)[2]
-                                                      
+
 function assemble_vrj(js, vrj, statetoid)
     rate   = generate_rate_function(js, vrj.rate)
     outputvars = (convert(Variable,affect.lhs) for affect in vrj.affect!)
@@ -90,19 +90,20 @@ function assemble_crj(js, crj, statetoid)
     ConstantRateJump(rate, affect)
 end
 
-function assemble_maj(js, maj::MassActionJump{U,Vector{Pair{V,W}},Vector{Pair{V2,W2}}},
-                      statetoid, subber, invttype) where {U,V,W,V2,W2}
-    sr = maj.scaled_rates
-    if sr isa Operation
-        pval = subber(sr).value
-    elseif sr isa Variable
-        pval = subber(sr()).value
+function numericrate(rate, subber)
+    if rate isa Operation
+        rval = subber(rate).value
+    elseif rate isa Variable
+        rval = subber(rate()).value
     else
-        pval = maj.scaled_rates
+        rval = rate
     end
+    rval
+end
 
+function numericrstoich(mtrs::Vector{Pair{V,W}}, statetoid) where {V,W}
     rs = Vector{Pair{Int,W}}()
-    for (spec,stoich) in maj.reactant_stoch
+    for (spec,stoich) in mtrs
         if iszero(spec)
             push!(rs, 0 => stoich)
         else
@@ -110,18 +111,37 @@ function assemble_maj(js, maj::MassActionJump{U,Vector{Pair{V,W}},Vector{Pair{V2
         end
     end
     sort!(rs)
+    rs
+end
 
-    ns = Vector{Pair{Int,W2}}()
-    for (spec,stoich) in maj.net_stoch
+function numericnstoich(mtrs::Vector{Pair{V,W}}, statetoid) where {V,W}
+    ns = Vector{Pair{Int,W}}()
+    for (spec,stoich) in mtrs
         iszero(spec) && error("Net stoichiometry can not have a species labelled 0.")
         push!(ns, statetoid[convert(Variable,spec)] => stoich)
     end
     sort!(ns)
-
-    maj = MassActionJump(convert(invttype, pval), rs, ns, scale_rates = false)
-    return maj
 end
 
+# assemble a numeric MassActionJump from a MT MassActionJump representing one rx.
+function assemble_maj(maj::MassActionJump{U,V,W}, statetoid, subber, 
+                      invttype) where {U,V,W}
+    rval = numericrate(maj.scaled_rates, subber)
+    rs   = numericrstoich(maj.reactant_stoch, statetoid)
+    ns   = numericnstoich(maj.net_stoch, statetoid)
+    maj  = MassActionJump(convert(invttype, rval), rs, ns, scale_rates = false)
+    maj
+end
+
+# For MassActionJumps that contain many reactions
+function assemble_maj(maj::MassActionJump{U,V,W}, statetoid, subber, 
+                      invttype) where {U <: AbstractVector,V,W}
+    rval = [convert(invttype,numericrate(sr, subber)) for sr in maj.scaled_rates]
+    rs   = [numericrstoich(rs, statetoid) for rs in maj.reactant_stoch]
+    ns   = [numericnstoich(ns, statetoid) for ns in maj.net_stoch]
+    maj  = MassActionJump(rval, rs, ns, scale_rates = false)
+    maj
+end
 """
 ```julia
 function DiffEqBase.DiscreteProblem(sys::AbstractSystem, u0map, tspan,
@@ -147,7 +167,7 @@ function DiffEqBase.DiscreteProblem(sys::AbstractSystem, u0map, tspan::Tuple,
     u0 = [u0d[u] for u in states(sys)]
     if parammap != DiffEqBase.NullParameters()
         (parammap isa AbstractVector) || error("For DiscreteProblems parammap must be an AbstractVector.")
-        pd  = Dict( convert(Variable,u[1]) => u[2] for u in parammap)    
+        pd  = Dict( convert(Variable,u[1]) => u[2] for u in parammap)
         p  = [pd[u] for u in parameters(sys)]
     else
         p = parammap
@@ -173,7 +193,7 @@ sol = solve(jprob, SSAStepper())
 """
 function DiffEqJump.JumpProblem(js::JumpSystem, prob, aggregator; kwargs...)
 
-    statetoid = Dict(convert(Variable,state) => i for (i,state) in enumerate(states(js)))    
+    statetoid = Dict(convert(Variable,state) => i for (i,state) in enumerate(states(js)))
     eqs       = equations(js)
     invttype  = typeof(1 / prob.tspan[2])
 
@@ -181,11 +201,11 @@ function DiffEqJump.JumpProblem(js::JumpSystem, prob, aggregator; kwargs...)
     p = (prob.p == DiffEqBase.NullParameters()) ? Operation[] : prob.p
     parammap  = map((x,y)->Pair(x(),y), parameters(js), p)
     subber    = substituter(first.(parammap), last.(parammap))
-    
-    majs = MassActionJump[assemble_maj(js, j, statetoid, subber, invttype) for j in eqs.x[1]]
+
+    majs = MassActionJump[assemble_maj(j, statetoid, subber, invttype) for j in eqs.x[1]]
     crjs = ConstantRateJump[assemble_crj(js, j, statetoid) for j in eqs.x[2]]
     vrjs = VariableRateJump[assemble_vrj(js, j, statetoid) for j in eqs.x[3]]
-    ((prob isa DiscreteProblem) && !isempty(vrjs)) && error("Use continuous problems such as an ODEProblem or a SDEProblem with VariableRateJumps") 
+    ((prob isa DiscreteProblem) && !isempty(vrjs)) && error("Use continuous problems such as an ODEProblem or a SDEProblem with VariableRateJumps")
     jset = JumpSet(Tuple(vrjs), Tuple(crjs), nothing, isempty(majs) ? nothing : majs)
 
     if needs_vartojumps_map(aggregator) || needs_depgraph(aggregator)
@@ -216,7 +236,7 @@ end
 ### Functions to determine which states are modified by a given jump
 function modified_states!(mstates, jump::Union{ConstantRateJump,VariableRateJump}, sts)
     for eq in jump.affect!
-        st = eq.lhs 
+        st = eq.lhs
         (st.op in sts) && push!(mstates, st)
     end
 end