To further improve large system performance, motivated by 
https://github.com/SciML/ModelingToolkit.jl/issues/426

and based on profiling the conversion and jump generation parts.