@isaacsas 2020-06-22
This speeds up the creation of `ReactionSystem`s and conversion to `JumpSystem`s a fair amount. I believe they are now allocation limited in performance, so we would require a more substantive alteration for further significant performance improvements. (Such as directly generating `JumpProblem`s, or at least the `MassActionJump`s, bypassing conversion.) 

The generation of `JumpProblem`s from a `JumpSystem` is the last piece to optimize.