@isaacsas 2020-06-28
I'm still not happy with where we are at, but probably this work should get merged. If I can get some time this week I'll try to cleanup and submit what is done -- the past few weeks have just been super busy... I've gotten very good speed ups on generating `ReactionSystem`s and `DiscreteProblem`s, and good speed ups on conversion to a `JumpSystem`. I haven't had a chance to work much on the `JumpSystem` to `JumpProblem` conversion, which is now the slowest component. 

More generally, we may want to rethink the internal design here somewhat. There are a lot of (intentional) allocations going on with having three levels of representations. This is one of two broad issues, the other being dynamic dispatch and instabilities due to using non-concrete types. For example, I believe conversion to `JumpSystem`s is now allocation limited in this PR. Profiling has been a pain here though, with many (to me) non-intuitive results. For example, in `ismassaction` I was seeing dynamic dispatch when using `any` (I think I've now eliminated that by manually searching), and significant time being spent in calls to `isempty` on an array. The latter is better when explicitly checking if the length is zero, but still a non-trivial amount of time, which just seems perplexing...

Ultimately, we may want to change JumpSystems to store `MassActionJumps` using integer identifiers for species, so we can reuse stoichiometry vectors in the `JumProblem`s, or perhaps avoid converting reaction systems to a `JumpSystem` intermediate. These would entail some larger changes though.