@isaacsas 2020-06-28
Along those lines; `Operation`s are not concretely typed right (given that they store abstract types, `Function`s and `Expression`s?)