@ChrisRackauckas 2020-06-28
>More generally, we may want to rethink the internal design here somewhat. There are a lot of (intentional) allocations going on with having three levels of representations. This is one of two broad issues, the other being dynamic dispatch and instabilities due to using non-concrete types. For example, I believe conversion to JumpSystems is now allocation limited in this PR. Profiling has been a pain here though, with many (to me) non-intuitive results. For example, in ismassaction I was seeing dynamic dispatch when using any (I think I've now eliminated that by manually searching), and significant time being spent in calls to isempty on an array. The latter is better when explicitly checking if the length is zero, but still a non-trivial amount of time, which just seems perplexing...

Yeah in terms of speed there's a lot to discuss. I've been discussing it with @YingboMa and @shashi in some other channels, but let's start taking it to #diffeq-bridged. There's a lot going on here. Essentially, it doesn't make sense to strictly type these because the structure of the equations is different every single time. If we concretely type the operations, then the type tree would encode the structure of the equations and therefore the type-stable functions for building the output would have to recompile for every new equation. So you'd make the compiled code fast but then you'd have to compile every time, which defeats the purpose of making it fast. So these pieces are kept abstractly typed so that it never has to recompile (I noticed this might be an issue in this PR). This BTW is the reason why most compilers, including Julia's, are not type-stable.

Thus making this code fast needs a different focus from standard Julia code. Specifically:

- One of the things that we need to think about for example is avoiding multiple dispatch since that's more expensive than a branch, so things like `assemble_crj` and `assemble_maj` are correctly doing the if statements to then call single dispatch functions. We need to find other dispatch hotspots and make them do case handling.
- We might want to collapse the types that we support to increase the number of single dispatch functions. Do we need `Constant(1)`? That's equivalent to `Operation(identity,Expression[1])`, so could we always use `Operation`, or is it lighter to not have that array? 
- Decrease the number of abstract types around. `Expression[x,y,z]` tends to give the compiler more trouble than just `Any[x,y,z]`, so we might want to just change `Operation` to the latter and then we might have more things resolve correctly in the compiler instead of relying on unions.
- Possibly use `(module,name)` (both as symbols) instead of `Function` inside of `Operation`. 

Etc. We should probably get an issue going to track updates in this as well.