@isaacsas 2020-06-29
Yeah, it would be great to be able to see that conversation through the Slack or issues. Probably much of what you discuss could be helpful for this PR too...

As I said, I'll try to cleanup and then update to master this week to merge what I have so far.