@isaacsas 2020-07-06
Timings on the SIR network model with 20000 edges:
```
PR:
julia> @time rs = ReactionSystem(rxs, t, vars, params);
  0.008907 seconds (14 allocations: 3.663 MiB)

julia> @time rs = ReactionSystem(rxs, t, vars, params);
  0.011149 seconds (14 allocations: 3.663 MiB)

Master: 
julia> @time rs = ReactionSystem(rxs, t, vars, params);
  0.302977 seconds (604.29 k allocations: 29.668 MiB)

julia> @time rs = ReactionSystem(rxs, t, vars, params);
  0.048388 seconds (179.00 k allocations: 6.394 MiB)

————————————————————————

PR:
julia> @time js = convert(JumpSystem, rs);
  0.371522 seconds (2.94 M allocations: 130.833 MiB)

julia> @time js = convert(JumpSystem, rs);
  0.411264 seconds (2.94 M allocations: 130.833 MiB)

Master:
julia> @time js = convert(JumpSystem, rs);
 60.578009 seconds (9.02 M allocations: 383.179 MiB, 0.27% gc time)

julia> @time js = convert(JumpSystem, rs);
 59.042692 seconds (5.68 M allocations: 213.643 MiB, 0.44% gc time)

————————————————————————

PR:
julia> @time dprob = DiscreteProblem(js, initial_state, tspan, initial_params)
  0.591422 seconds (1.29 M allocations: 67.144 MiB)

julia> @time dprob = DiscreteProblem(js, initial_state, tspan, initial_params)
  0.202933 seconds (597.07 k allocations: 30.512 MiB)

Master:
julia> @time dprob = DiscreteProblem(js, initial_state, tspan, initial_params)
 44.106209 seconds (1.52 M allocations: 86.365 MiB)

julia> @time dprob = DiscreteProblem(js, initial_state, tspan, initial_params)
 43.070374 seconds (239.78 k allocations: 13.723 MiB)

————————————————————————
PR:
julia> @time jprob = JumpProblem(js, dprob, Direct(), save_positions=(false,false))
  3.000220 seconds (9.75 M allocations: 277.510 MiB, 23.84% gc time)

julia> @time jprob = JumpProblem(js, dprob, Direct(), save_positions=(false,false))
  2.185403 seconds (9.35 M allocations: 255.276 MiB, 6.34% gc time)

Master:
julia> @time jprob = JumpProblem(js, dprob, Direct(), save_positions=(false,false))
 12.597332 seconds (56.60 M allocations: 1.456 GiB, 11.01% gc time)

julia> @time jprob = JumpProblem(js, dprob, Direct(), save_positions=(false,false))
  9.869804 seconds (50.79 M allocations: 1.159 GiB, 13.19% gc time)
```