@isaacsas 2020-07-06
@ChrisRackauckas Let me know if you have any comments. Otherwise, if tests pass I'd say to merge and I'll keep working on this as I have time. There is definitely more to do, especially with reducing how many times `MassActionJump`s get created going between `ReactionSystem`s, `JumpSystem`s and the final `MassActionJump` for the solvers.